+++ packages/scripts/addon.js
@@ -1,17 +1,16 @@
 #!/usr/bin/env node
 /* eslint no-console: 0 */
 /**
- * testing-addon script.
- * @module scripts/testing-addon
+ * addon testing setup convenience script
+ * @module scripts/addon
  */
 const { program } = require('commander');
-const { exec } = require('child_process');
+const { execSync } = require('child_process');
 const https = require('https');
 const GitUrlParse = require('git-url-parse');
 const fs = require('fs');
 const { develop } = require('mrs-developer');
-
-function main() {}
+const chalk = require('chalk');
 
 function amendPackageJSON(name, destination) {
   const packageJSON = JSON.parse(
@@ -21,7 +20,6 @@ function amendPackageJSON(name, destination) {
     ...packageJSON.scripts,
     'cypress:open': `cd src/addons/${name} && NODE_ENV=test cypress open`,
   };
-  console.log(packageJSON);
   fs.writeFileSync(
     `${destination}/package.json`,
     `${JSON.stringify(packageJSON, null, 4)}`,
@@ -33,6 +31,7 @@ function createMrsDeveloperConfig(config) {
   const packageName = config.fullname || config.name;
   template[config.name].package = packageName;
   template[config.name].url = config.source;
+  template[config.name].path = 'src';
   template[config.name].branch = config.branch || 'main';
 
   fs.writeFileSync(
@@ -44,50 +43,89 @@ function createMrsDeveloperConfig(config) {
 /*
  * Retrieves latest Volto released version from NPM registry
  */
-async function getAddonInfo(source, branch = 'main') {
+async function getAddonInfo({ source, branch = 'main', isPrivate }) {
   const sourceParse = GitUrlParse(source);
-  const httpRawURL = sourceParse
-    .toString('https')
-    .replace('github.com', 'raw.githubusercontent.com');
-  const packageJSONURL = `${httpRawURL}/${branch}/package.json`;
-
-  console.log(sourceParse);
-  console.log(packageJSONURL);
+  const packageJSONURL = isPrivate
+    ? `https://api.github.com/repos/${sourceParse.full_name}/contents/package.json`
+    : `https://raw.githubusercontent.com/${sourceParse.full_name}/${branch}/package.json`;
 
   return new Promise((resolve, reject) => {
     https
-      .get(packageJSONURL, { headers: {} }, (resp) => {
-        let data = [];
-        resp.on('data', (chunk) => {
-          data.push(chunk);
-        });
-        resp.on('end', () => {
-          const res = JSON.parse(data.join(''));
-          resolve({
-            name: res.name,
-            version: res.version,
-            ...(res.name.includes('@')
-              ? { fullname: res.name, name: res.name.split('/')[1] }
+      .get(
+        packageJSONURL,
+        {
+          headers: {
+            ...(isPrivate
+              ? {
+                  Authorization: `token ${process.env.GITHUB_TOKEN}`,
+                  'User-Agent': 'node.js',
+                  Accept: 'application/vnd.github.VERSION.raw',
+                }
               : {}),
+          },
+        },
+        (resp) => {
+          let data = [];
+          resp.on('data', (chunk) => {
+            data.push(chunk);
           });
-        });
-      })
+          resp.on('end', () => {
+            const res = JSON.parse(data.join(''));
+            resolve({
+              name: res.name,
+              version: res.version,
+              ...(res.name.includes('@')
+                ? { fullname: res.name, name: res.name.split('/')[1] }
+                : {}),
+            });
+          });
+        },
+      )
       .on('error', (err) => {
         reject(err.message);
       });
   });
 }
 
-async function runGenerator(source, destination = 'addon-testing-project') {
-  const { fullname, name, version } = await getAddonInfo(source);
-  console.log(fullname);
-  console.log(name);
-  console.log(version);
+async function runGenerator({
+  source,
+  destination = 'addon-testing-project',
+  isPrivate = false,
+  branch = 'main',
+}) {
+  if (isPrivate) {
+    console.log(
+      chalk.green(
+        `Retrieving private package info from branch ${chalk.yellow(
+          branch,
+        )} repo ${chalk.yellow(source)}`,
+      ),
+    );
+  } else {
+    console.log(
+      chalk.blue(
+        `Retrieving package info from branch ${chalk.yellow(
+          branch,
+        )} repo ${chalk.yellow(source)}`,
+      ),
+    );
+  }
+
+  const { fullname, name, version } = await getAddonInfo({ source, isPrivate });
+
+  console.log(
+    chalk.green(
+      `The package name is ${chalk.yellow(
+        fullname || name,
+      )} pulling version ${chalk.yellow(version)}`,
+    ),
+  );
+
   const GENERATOR_CLI = `yo --force --no-insight @plone/volto ${destination} --no-interactive --skip-install --workspace src/addons/${name} --addon ${
     fullname || name
   }`;
 
-  exec(GENERATOR_CLI, (error, stdout, stderr) => {
+  execSync(GENERATOR_CLI, (error, stdout, stderr) => {
     if (error) {
       console.log(`error: ${error.message}`);
       return;
@@ -96,9 +134,17 @@ async function runGenerator(source, destination = 'addon-testing-project') {
       console.log(`stderr: ${stderr}`);
       return;
     }
-    console.log(`stdout: ${stdout}`);
+    //  console.log(`stdout: ${stdout}`);
   });
-  createMrsDeveloperConfig({ fullname, name, version, source, destination });
+
+  createMrsDeveloperConfig({
+    fullname,
+    name,
+    version,
+    source,
+    destination,
+    branch,
+  });
 
   await develop({
     root: destination,
@@ -106,7 +152,7 @@ async function runGenerator(source, destination = 'addon-testing-project') {
     output: 'addons',
   });
 
-  exec(`cd ${destination} && yarn`, (error, stdout, stderr) => {
+  execSync(`cd ${destination} && yarn`, (error, stdout, stderr) => {
     if (error) {
       console.log(`error: ${error.message}`);
       return;
@@ -115,16 +161,31 @@ async function runGenerator(source, destination = 'addon-testing-project') {
       console.log(`stderr: ${stderr}`);
       return;
     }
-    console.log(`stdout: ${stdout}`);
+    //  console.log(`stdout: ${stdout}`);
   });
 
+  console.log(
+    chalk.green(
+      `Preparing and amending testing project ${chalk.yellow('package.json')}`,
+    ),
+  );
   amendPackageJSON(name, destination);
 }
 
-function cloneAddon(source, destination) {
-  console.log(`${source} : ${destination}`);
-  //  exec();
-  runGenerator(source, destination).then();
+function cloneAddon({
+  source,
+  destination = 'addon-testing-project',
+  branch = 'main',
+  isPrivate,
+}) {
+  console.log(
+    chalk.green(
+      `Cloning addon from ${chalk.yellow(
+        source,
+      )} and creating a testing environment in ${chalk.yellow(destination)}`,
+    ),
+  );
+  runGenerator({ source, destination, branch, isPrivate }).then();
 }
 
 // This is the equivalent of `if __name__ == '__main__'` in Python :)
@@ -132,10 +193,18 @@ if (require.main === module) {
   program
     .command('clone <source> [destination]')
     .description('clone a repository into a newly created directory')
-    .action((source, destination) => {
-      cloneAddon(source, destination);
+    .option(
+      '-p, --private',
+      'set if the repo is private, then GITHUB_TOKEN is used',
+    )
+    .option('-b, --branch <branch>', 'set the repo branch, defaults to main')
+    .action((source, destination, options) => {
+      cloneAddon({
+        source,
+        destination,
+        isPrivate: options.private,
+        branch: options.branch,
+      });
     });
   program.parse(process.argv);
-  const options = program.opts();
-  main({ addonMode: options.addon });
 }